Test project, fix directory structure and imports.
In main there is an example on how to run a fault injection
with standard-like settings.

The directory structure has been slightly tweaked to have all
the model-related stuff under dnn.

Imports are more pythonic, requiring no knowledge of the
principal directory.

Some small fixes have been introduced, regarding the way
BaseFI sets up its own configuration, using a method init instead
of the object initializer __init__.
LeNet5 implementation is closer to the one in PyTorch and in the paper.