Implement input/output size and model summary.
BaseModule now supports both a staticmethod for returning the model
summary as well as a property for the summary of the module itself.
This summary uses torch-summary from pip to gather information on
model size and module outputs, which are required when injecting onto
the output of a module.
Moreover it supports input and output sizes, allowing customization
of standard models onto different datasets.