Base structure, must be updated.
The current structure requires an update, as we need to include
also a timeline and a mapping for each kernel that needs to be run.

The solution is to have a timeline model, where we have a set of kernels running for a certain amount of time, therefore we know exactly which kernels are struck by the particles. Then, each kernel has its own fault mapping, depending on where the chip is hit. Each of them will have a probability of having a fault in a certain bit position depending on the original fault. This probability/distribution can be adapted depending on different kernels/hardware, by doing some manual tests. Otherwise, a future solution will be to learn these distributions by using different DNN models to have a sort of digital twin of the hardware.