Adjust component count and scheduling signature.
Now the components are instantiated using
number_of_components_per_parent, to be explicit about counting the
number of components per parent instance and not in absolute,
to make modeling easier.

Improved the arguments for scheduling an inference run, as it does not
require the target IDs anymore. In this way, we assume the target
components are numbered from 0 to n_components - 1, and they are
all free when starting the current run. While this may be limiting,
it should be enough for the current experiments, and it can be extended
in the future.

Insert the current changes in the proper docs.