Fix tests and implement new JSON converter.
Tests have been slightly expanded to cover also the new functions.

Bugs found from the tests have also been fixed, mostly related to name
typos.

Improve the way we handling JSON conversion, now we have a base class
that registers via decorator the encoder and decoder for that specific
class.

Moreover, also the current sample hierarchy has been updated,
together with the corresponding parsing code in the HardwareModel.
The change has required moving from typing.NamedTuple to
dataclasses.dataclass for GPUComponentHierarchy and GPUComponent,
as typing.NamedTuple does not support inheritance for allowing the base
JSONSerializable* class usage as default.

Add also method for returning the stripped JSON for the HardwareModel
hierarchy, as the internal model uses the enum as key for a dict for
lookup efficiency. This property method as well as the other expensive
ones (hierarchy and map) have been converted to
functools.cached_property, to improve performance for some memory
expense.

Update also the DESIGN_CHOICES document with the current decisions.