Update structure, implement activation injection.
Extend test to cover weight and activation injection.

Fix wrong pre/post function names for callback.

Add enabling/disabling single faults in InjectionCallback.

Implement Endianness for bit ordering in FaultDescriptor.

Implement bit and tensor index parsing in FaultDescriptor.

Make FaultDescriptor hashable with slices for use as dict key.

Update ParameterType to be Flag-based and not a simple Enum.
This eases the operations of comparing to which type the
parameter belongs to.

Implement post-init checks for InjectionCallback and FaultDescriptor.

Implement activation injection as torch.nn.Module.