Update weight injection to use masks & converters.
Update some usages of class.__name__ to class.__qualname__,
which is supported from Python 3.3 and better for class names,
to cover for recursive classes.

Define fill_value in BinaryMask, which depends on the operation.

Define converter for CuPy.

Add conversion in type for bitwise operations in NumPy
converter, as well as mask bit expansion to whole numpy.array.

Comment and expand PyTorch converter to cover also CuPy.

Add some extra functionalities for PyTorch converter, regarding
device and device type conversion.

Implement PyTorch device-aware converter, which uses the Dispatcher to dispatch the calls to the correct handler, either NumPy or CuPy depending on the device which is used.

Fix bugs with endianness, list instead of string for setting values
and StuckAtZero error for binary injection handler.