Add versioning and time info in SaveConfigCallback
Now in the configuration one can specify whether versioning
should be done inside the configs directory.
Depending on the argument, it will be treated as the version
number or the latest version number incremented by one will be used.
In addition also a time signature can be used.

The basic config has been updated to reflect this.