Implement JSONNET configs for Flash CLI.
There are still bugs in Flash CLI, as it requires a sequential argument
to toggle the DataModule instantiation,
but it gets stuck to the default one when using a config file.

Additionally it has still issues parsing nested
class_path instantiations.

So it will not be used in the near future.