feat(plugins): indexing plugin and related updates
IndexingPlugin allows for custom indexing when injecting faults.

In this way, the faults from the tensors are selected properly depending
on the default indexing that is used and the selected active dimensions.

Include new enumeration type, indexing constants for PyTorch and Norse,
typings.

Fix errors related to EllipsisType for Python < 3.10.