fix(mask): fix mask dimension if injecting weights
Mask dimension would happen to be wrong when injecting weights,
as they would have no batch size to remove at the start.

The fix is to add another parameter to indicate whether
the batch dimension exists in the target.

The appropriate indexing function has been adapted
as well to accomodate the weight injection.

The default behaviour stays identical to the previous one.