feat(ijcnn2022): improve image classification conf
Add support for automatic numpy/cupy switch.

Add arguments for model weight and storage file.

Fix faults to use Weight fault if happening on weights.
This speeds up execution as the weights are changed only once.

Fine-tune faults to slightly reduce accuracy instead of
flattening it to random behaviour.

Instantiate callback as late as possible,
so the config can be saved as extra session info,
since it does not contain any instantiated elements.