feat: add parallel injection script
Improve on the injection script, and add a parallel injection script
for multi-GPU systems, with full configurability.

Add also plotting notebooks with temporary results, as well as new
training scripts for possible future results.

There is also a placeholder for a helper containing things related
to Importance Sampling, which will be used later.